Fixes broken test; removes some duplicate code & cleans up types

+++ grails-app/controllers/au/org/ala/profile/hub/ProfileController.groovy
@@ -313,7 +313,7 @@ class ProfileController extends BaseController {
         } else {
             boolean latest = params.isOpusReviewer || params.isOpusEditor || params.isOpusAdmin
             boolean readonlyView = params.getBoolean('readonlyView', true)
-            def  response = imageService.retrieveImagesPaged(params.opusId, params.profileId, latest, params.searchIdentifier, false, readonlyView, params.pageSize, params.startIndex)
+            def  response = imageService.retrieveImagesPaged(params.opusId, params.profileId, latest, params.searchIdentifier, false, readonlyView, params.pageSize as int, params.startIndex as int)
             handle (response)
         }
     }
